move games/screensavers to SD card
