--- src/MicroPython/picoware/applications/games/games.py
@@ -1,81 +1,0 @@
-_games = None
-_games_index = 0
-
-
-def start(view_manager) -> bool:
-    """Start the app"""
-    from picoware.gui.menu import Menu
-
-    global _games
-    global _games_index
-    if _games is None:
-        _games = Menu(
-            view_manager.draw,
-            "Games",
-            0,
-            320,
-            view_manager.get_foreground_color(),
-            view_manager.get_background_color(),
-            view_manager.get_selected_color(),
-            view_manager.get_foreground_color(),
-            2,
-        )
-        _games.add_item("Example")
-        _games.set_selected(_games_index)
-
-        _games.draw()
-    return True
-
-
-def run(view_manager) -> None:
-    """Run the app."""
-    from picoware.system.view import View
-    from picoware.system.buttons import (
-        BUTTON_BACK,
-        BUTTON_UP,
-        BUTTON_DOWN,
-        BUTTON_LEFT,
-        BUTTON_CENTER,
-        BUTTON_RIGHT,
-    )
-
-    global _games
-    if not _games:
-        return
-    global _games_index
-
-    input_manager = view_manager.input_manager
-    button: int = input_manager.get_last_button()
-
-    if button == BUTTON_UP:
-        input_manager.reset(True)
-        _games.scroll_up()
-    elif button == BUTTON_DOWN:
-        input_manager.reset(True)
-        _games.scroll_down()
-    elif button in (BUTTON_BACK, BUTTON_LEFT):
-        _games_index = 0
-        input_manager.reset(True)
-        view_manager.back()
-    elif button in (BUTTON_CENTER, BUTTON_RIGHT):
-        input_manager.reset(True)
-        _games_index = _games.get_selected_index()
-
-        if _games_index == 0:
-            from picoware.applications.games import example
-
-            view_manager.add(
-                View("games_example", example.run, example.start, example.stop)
-            )
-            view_manager.switch_to("games_example")
-
-
-def stop(view_manager) -> None:
-    """Stop the app"""
-    from gc import collect
-
-    global _games
-    if _games is not None:
-        del _games
-        _games = None
-    collect()
